+++ modules/variables.js
@@ -1,0 +1,15 @@
+// for display.js
+export const error = document.getElementById('error');
+export const displayBookList = document.getElementById('table');
+export const bookTitle = document.getElementById('book-title');
+export const author = document.getElementById('author');
+// for index.js
+export const submitButton = document.getElementById('add');
+export const listLink = document.getElementById('list-link');
+export const addNewLink = document.getElementById('add-new-link');
+export const contactLink = document.getElementById('contact-link');
+export const form = document.getElementById('form');
+// for both
+export const listSection = document.getElementById('list');
+export const newBookSection = document.getElementById('book-form');
+export const contactSection = document.getElementById('contact-info');
